New logging macros
This commit is only to introduce new logging macros.

The current implementation is such that a statement like:

    MAXSCALE_NOTICE("Refreshing configuration following SIGHUP\n");

is equivalent with

    LOGIF(LM, (skygw_log_write(
                LOGFILE_MESSAGE,
                "Refreshing configuration following SIGHUP\n")));

The actual implementation will later be changed as the logging
mechanism itself is changed.

The names of the macros are now according to the levels of syslog
and currently the mapping is like:

    MAXSCALE_ERROR   (Syslog LOG_ERR)     -> LOGFILE_ERROR
    MAXSCALE_WARNING (Syslog LOG_WARNING) -> LOGFILE_ERROR
    MAXSCALE_NOTICE  (Syslof LOG_NOTICE)  -> LOGFILE_MESSAGE
    MAXSCALE_INFO    (Syslog LOG_INFO)    -> LOGFILE_TRACE
    MAXSCALE_DEBUG   (Syslog LOG_DEBUG)   -> LOGFILE_DEBUG

When log manager is changed to deal "natively" with syslog levels
this mapping will disappear of course.