Fix to MXS-367: https://mariadb.atlassian.net/browse/MXS-367
Template config file is copied as maxscale.cnf if no configuration file is found.
If /etc/maxscale.cnf is not found, the /etc/maxscale.cnf.template will be copied as /etc/maxscale.cnf.
This allows MaxScale to start right after installation if combined with a running and properly configured server.