Removed redundant information from Galera tutorials.
The Galera tutorials were very similar to the MySQL Replication ones with the
only difference being the configuration file creation. The tutorials were split
into smaller parts which are now referred to in the main MaxScale tutorial.

Renamed MaxScale-and-Master-Slave-Replication.md to MaxScale-Tutorial.md since
the tutorial wasn't only about Master-Slave replication and included the links
to the Galera tutorials.