Fixed postinstall script
The postinstall script tried to create a group for the maxscale even if the
useradd command had created one. This fix removes the warning messages by
creating the group while the user is being created.

If the systemd folder wasn't located at /usr/lib/systemd/ but at /lib/systemd/
the postinstall script would not copy the file.

Also added error messages if some of the files could not be located.