server_weight falsely assumed that the first node is always the master
The server_weight test used a Galera cluster which does not always have
the same node as the master. The second node was usually chosed as the master
which caused the test to fail.

With the `priority` mechanism provided by the galeramon, we can force the first
node as the master and we can safely assume that the replication topology is
always the same.