Update CMakeLists.txt for qc_sqlite
Library is no longer linked with "-Wl,-z,defs", which would cause
all symbols to be resolved. That would need qc_sqlite to be linked
with libmaxscale-common and for whatever reason that leads to
sqlite3 being used from the system libsqlite3 and not from the
sqlite3 built into libqc_sqlite.

Now also sqlite3 is built using the compiler commands of MaxScale
so that if MaxScale is built in debug mode, so is sqlite3. As long
as libqc_sqlite is being developed that makes things easier.