Re-route sqlite's sqlite3Select.
The parser of sqlite calls sqlite3Select when it has parsed a valid
select statement. The function then generates code for sqlite's virtual
machine.

That function is now replaced with a MaxScale specific one that later
will collect the information we are interested in and return an error
that aborts sqlite's processing.