Add minimal select recognition to qc_sqlite
This change now implements in principle the basic mechanism of the
sqlite based query classifier.

When a information about a query residing in a GWBUF is asked for,
the procedure (if the query has not been parsed yet) is as follows
(see parse_query):

- Allocate a QC_SQLITE_INFO object for holding information about
  the query. The status is set to QC_INFO_INVALID.
- Store a pointer to the info object in the thread local struct
  this_thread.
- Call sqlite3_prepare(). That'll parse the query and in the end one
  of our replacement functions will be called (currently only
  qc_sqlite3Select).
- In that function we collect information about the query and
  set the info status to QC_INFO_OK.
- That is, when sqlite3_prepare() returns, if the status field in
  the info object is QC_INFO_OK, we know the call chain has
  reached one of our "overriding" functions and that we have
  collected information about the query.
- The info object, valid or not, is always added to the GWBUF.
  If the parsing failed this time, it will fail next time as well.
- Thereafter, if more information is asked about the query, we can
  consult the info object directly.

The following step is to create a test-suite for comparing the
behaviour of the MySQL based query classifier (qc_mysqlembedded)
and this, and then fix discrepancies.