MXS-633: Monitor permission checks moved to modules
Each of the monitors require different privileges so either the modules
need to declare the permissions or do the checks themselves. This change
moved the checks to monitor modules and makes the permission checking function
more easy to use.

In the future modules could declare more information about what they need and
what they provide. Until that mechanism is implemented, the modules have to
do the work.