Cleaned up core test suite
The tests can now be run without requiring MaxScale to be build with
special flags. This allows running of the `make test` target to make
sure that the compilation was successful before installing the files.

The cleaning up of the test suite meant that all tests which require
external dependencies were removed. Very similar tests are executed
as a part of the more complex system test suite.