Removed messages about incomplete transaction events
The message is logged when a DDL statement is executed. It should not be
logged if trx_safe is on since the current_safe_event should always point
at the event we are sending. The current_safe_event is set to the wrong value
which causes this message to be logged.

Due to the false positives caused by this, the message is removed.